Updated: Integer -> Short in TabCommand

+++ modules/RoyalCommands/src/main/java/org/royaldev/royalcommands/rcommands/CmdAssign.java
@@ -17,7 +17,7 @@ import java.util.List;
 public class CmdAssign extends TabCommand {
 
     public CmdAssign(final RoyalCommands instance, final String name) {
-        super(instance, name, true, new Integer[]{CompletionType.ANY_COMMAND.getInt()});
+        super(instance, name, true, new Short[]{CompletionType.ANY_COMMAND.getShort()});
     }
 
     @Override
