+++ modules/RoyalCommands/src/main/java/org/royaldev/royalcommands/rcommands/CmdUses.java
@@ -34,7 +34,7 @@ public class CmdUses extends TabCommand {
     private final Map<String, Integer> tasks = new HashMap<>();
 
     public CmdUses(final RoyalCommands instance, final String name) {
-        super(instance, name, true, new Integer[]{CompletionType.ITEM_ALIAS.getInt()});
+        super(instance, name, true, new Short[]{CompletionType.ITEM_ALIAS.getShort()});
         this.plugin.getServer().getPluginManager().registerEvents(new WorkbenchCloseListener(), this.plugin);
     }
